services/driver-monitor/requirements.txt
redis>=4.0.0
requests>=2.28.0
pyyaml>=6.0
apscheduler>=3.10.0
mwdblib>=4.0.0
vt-py>=0.18.0
